at_depth {purrr}	R Documentation

Map at depth

Description

This function is defunct and has been replaced by map_depth(). See also modify_depth() for a version that preserves the types of the elements of the tree.

Usage

at_depth(.x, .depth, .f, ...)

[Package purrr version 1.0.4 Index]