auto_browse {purrr}	R Documentation

Wrap a function so it will automatically browse() on error

Description

A function wrapped with auto_browse() will automatically enter an interactive debugger using browser() when ever it encounters an error.

Usage

auto_browse(.f)

Arguments

.f
A function to modify, specified in one of the following ways:

A named function, e.g. mean.

An anonymous function, e.g. \(x) x + 1 or function(x) x + 1.

A formula, e.g. ~ .x + 1. Only recommended if you require backward compatibility with older versions of R.

Value

A function that takes the same arguments as .f, but returns a different value, as described above.

Adverbs

This function is called an adverb because it modifies the effect of a function (a verb). If you'd like to include a function created an adverb in a package, be sure to read faq-adverbs-export.

See Also

Other adverbs: compose(), insistently(), negate(), partial(), possibly(), quietly(), safely(), slowly()

Examples

# For interactive usage, auto_browse() is useful because it automatically
# starts a browser() in the right place.
f <- function(x) {
  y <- 20
  if (x > 5) {
    stop("!")
  } else {
    x
  }
}
if (interactive()) {
  map(1:6, auto_browse(f))
}

[Package purrr version 1.0.4 Index]