every {purrr}	R Documentation

Do every or some elements of a list satisfy a predicate?

Description

Do every or some elements of a list satisfy a predicate?

Usage

every(.x, .p, ...)

some(.x, .p, ...)

Arguments

.x
A list or atomic vector.

.p
A single predicate function, a formula describing such a predicate function, or a logical vector of the same length as .x. Alternatively, if the elements of .x are themselves lists of objects, a string indicating the name of a logical element in the inner lists. Only those elements where .p evaluates to TRUE will be modified.

...
Additional arguments passed on to .f.

Value

A logical vector of length 1.

Examples

x <- list(0, 1, TRUE)
x %>% every(identity)
x %>% some(identity)

y <- list(0:10, 5.5)
y %>% every(is.numeric)
y %>% every(is.integer)

[Package purrr version 0.2.5 Index]