Map/modify elements at given depth

map_depth() calls map(.y, .f) on all .y at the specified .depth in .x. modify_depth() calls modify(.y, .f) on .y at the specified .depth in .x.

Usage

map_depth(.x, .depth, .f, ..., .ragged = FALSE)

modify_depth(.x, .depth, .f, ..., .ragged = .depth < 0)

Arguments

.x

A list or atomic vector.

.depth

Level of .x to map on. Use a negative value to count up from the lowest level of the list.

map_depth(x, 0, fun) is equivalent to fun(x).

map_depth(x, 1, fun) is equivalent to x <- map(x, fun)

map_depth(x, 2, fun) is equivalent to x <- map(x, ~ map(., fun))

.f

A function, specified in one of the following ways:

A named function, e.g. mean.

An anonymous function, e.g. \(x) x + 1 or function(x) x + 1.

A formula, e.g. ~ .x + 1. You must use .x to refer to the first argument. Only recommended if you require backward compatibility with older versions of R.

A string, integer, or list, e.g. "idx", 1, or list("idx", 1) which are shorthand for \(x) pluck(x, "idx"), \(x) pluck(x, 1), and \(x) pluck(x, "idx", 1) respectively. Optionally supply .default to set a default value if the indexed element is NULL or does not exist.

...

Additional arguments passed on to the mapped function.

Note that the arguments that differ in each call come before .f, and the arguments that are the same come after .f.

.ragged

If TRUE, will apply to leaves, even if they're not at depth .depth. If FALSE, will throw an error if there are no elements at depth .depth.

Examples

# map_depth() -------------------------------------------------
# Use `map_depth()` to recursively traverse nested vectors and map
# a function at a certain depth:
x <- list(a = list(foo = 1:2, bar = 3:4), b = list(baz = 5:6))
str(x)
#> List of 2
#>  $ a:List of 2
#>   ..$ foo: int [1:2] 1 2
#>   ..$ bar: int [1:2] 3 4
#>  $ b:List of 1
#>   ..$ baz: int [1:2] 5 6
map_depth(x, 2, paste, collapse = "/")
#> $a
#> $a$foo
#> [1] "1/2"
#> 
#> $a$bar
#> [1] "3/4"
#> 
#> 
#> $b
#> $b$baz
#> [1] "5/6"
#> 
#> 

# Equivalent to:
map(x, map, paste, collapse = "/")
#> $a
#> $a$foo
#> [1] "1/2"
#> 
#> $a$bar
#> [1] "3/4"
#> 
#> 
#> $b
#> $b$baz
#> [1] "5/6"
#> 
#> 

# When ragged is TRUE, `.f()` will also be passed leaves at depth < `.depth`
x <- list(1, list(1, list(1, list(1, 1))))
str(x)
#> List of 2
#>  $ : num 1
#>  $ :List of 2
#>   ..$ : num 1
#>   ..$ :List of 2
#>   .. ..$ : num 1
#>   .. ..$ :List of 2
#>   .. .. ..$ : num 1
#>   .. .. ..$ : num 1
str(map_depth(x, 4, ~ length(unlist(.x)), .ragged = TRUE))
#> List of 2
#>  $ :List of 1
#>   ..$ : int 1
#>  $ :List of 2
#>   ..$ :List of 1
#>   .. ..$ : int 1
#>   ..$ :List of 2
#>   .. ..$ :List of 1
#>   .. .. ..$ : int 1
#>   .. ..$ :List of 2
#>   .. .. ..$ : int 1
#>   .. .. ..$ : int 1
str(map_depth(x, 3, ~ length(unlist(.x)), .ragged = TRUE))
#> List of 2
#>  $ :List of 1
#>   ..$ : int 1
#>  $ :List of 2
#>   ..$ :List of 1
#>   .. ..$ : int 1
#>   ..$ :List of 2
#>   .. ..$ : int 1
#>   .. ..$ : int 2
str(map_depth(x, 2, ~ length(unlist(.x)), .ragged = TRUE))
#> List of 2
#>  $ :List of 1
#>   ..$ : int 1
#>  $ :List of 2
#>   ..$ : int 1
#>   ..$ : int 3
str(map_depth(x, 1, ~ length(unlist(.x)), .ragged = TRUE))
#> List of 2
#>  $ : int 1
#>  $ : int 4
str(map_depth(x, 0, ~ length(unlist(.x)), .ragged = TRUE))
#>  int 5

# modify_depth() -------------------------------------------------
l1 <- list(
  obj1 = list(
    prop1 = list(param1 = 1:2, param2 = 3:4),
    prop2 = list(param1 = 5:6, param2 = 7:8)
  ),
  obj2 = list(
    prop1 = list(param1 = 9:10, param2 = 11:12),
    prop2 = list(param1 = 12:14, param2 = 15:17)
  )
)

# In the above list, "obj" is level 1, "prop" is level 2 and "param"
# is level 3. To apply sum() on all params, we map it at depth 3:
l1 %>% modify_depth(3, sum) %>% str()
#> List of 2
#>  $ obj1:List of 2
#>   ..$ prop1:List of 2
#>   .. ..$ param1: int 3
#>   .. ..$ param2: int 7
#>   ..$ prop2:List of 2
#>   .. ..$ param1: int 11
#>   .. ..$ param2: int 15
#>  $ obj2:List of 2
#>   ..$ prop1:List of 2
#>   .. ..$ param1: int 19
#>   .. ..$ param2: int 23
#>   ..$ prop2:List of 2
#>   .. ..$ param1: int 39
#>   .. ..$ param2: int 48

# Note that vectorised operations will yield the same result when
# applied at the list level as when applied at the atomic result.
# The former is more efficient because it takes advantage of
# vectorisation.
l1 %>% modify_depth(3, `+`, 100L)
#> $obj1
#> $obj1$prop1
#> $obj1$prop1$param1
#> [1] 101 102
#> 
#> $obj1$prop1$param2
#> [1] 103 104
#> 
#> 
#> $obj1$prop2
#> $obj1$prop2$param1
#> [1] 105 106
#> 
#> $obj1$prop2$param2
#> [1] 107 108
#> 
#> 
#> 
#> $obj2
#> $obj2$prop1
#> $obj2$prop1$param1
#> [1] 109 110
#> 
#> $obj2$prop1$param2
#> [1] 111 112
#> 
#> 
#> $obj2$prop2
#> $obj2$prop2$param1
#> [1] 112 113 114
#> 
#> $obj2$prop2$param2
#> [1] 115 116 117
#> 
#> 
#> 
l1 %>% modify_depth(4, `+`, 100L)
#> $obj1
#> $obj1$prop1
#> $obj1$prop1$param1
#> [1] 101 102
#> 
#> $obj1$prop1$param2
#> [1] 103 104
#> 
#> 
#> $obj1$prop2
#> $obj1$prop2$param1
#> [1] 105 106
#> 
#> $obj1$prop2$param2
#> [1] 107 108
#> 
#> 
#> 
#> $obj2
#> $obj2$prop1
#> $obj2$prop1$param1
#> [1] 109 110
#> 
#> $obj2$prop1$param2
#> [1] 111 112
#> 
#> 
#> $obj2$prop2
#> $obj2$prop2$param1
#> [1] 112 113 114
#> 
#> $obj2$prop2$param2
#> [1] 115 116 117
#> 
#> 
#> 

# modify() lets us pluck the elements prop1/param2 in obj1 and obj2:
l1 %>% modify(c("prop1", "param2")) %>% str()
#> List of 2
#>  $ obj1: int [1:2] 3 4
#>  $ obj2: int [1:2] 11 12

# But what if we want to pluck all param2 elements? Then we need to
# act at a lower level:
l1 %>% modify_depth(2, "param2") %>% str()
#> List of 2
#>  $ obj1:List of 2
#>   ..$ prop1: int [1:2] 3 4
#>   ..$ prop2: int [1:2] 7 8
#>  $ obj2:List of 2
#>   ..$ prop1: int [1:2] 11 12
#>   ..$ prop2: int [1:3] 15 16 17

# modify_depth() can be with other purrr functions to make them operate at
# a lower level. Here we ask pmap() to map paste() simultaneously over all
# elements of the objects at the second level. paste() is effectively
# mapped at level 3.
l1 %>% modify_depth(2, ~ pmap(., paste, sep = " / ")) %>% str()
#> List of 2
#>  $ obj1:List of 2
#>   ..$ prop1:List of 2
#>   .. ..$ : chr "1 / 3"
#>   .. ..$ : chr "2 / 4"
#>   ..$ prop2:List of 2
#>   .. ..$ : chr "5 / 7"
#>   .. ..$ : chr "6 / 8"
#>  $ obj2:List of 2
#>   ..$ prop1:List of 2
#>   .. ..$ : chr "9 / 11"
#>   .. ..$ : chr "10 / 12"
#>   ..$ prop2:List of 3
#>   .. ..$ : chr "12 / 15"
#>   .. ..$ : chr "13 / 16"
#>   .. ..$ : chr "14 / 17"