pluck_depth {purrr}	R Documentation

Compute the depth of a vector

Description

The depth of a vector is how many levels that you can index/pluck into it. pluck_depth() was previously called vec_depth().

Usage

pluck_depth(x, is_node = NULL)

Arguments

x
A vector

is_node
Optionally override the default criteria for determine an element can be recursed within. The default matches the behaviour of pluck() which can recurse into lists and expressions.

Value

An integer.

Examples

These examples are designed to work in R >= 4.1 so that we can take advantage of modern syntax like the base pipe (|>) and the function shorthand (\(x) x + 1). They might not work on the version of R that you're using.

x <- list(
  list(),
  list(list()),
  list(list(list(1)))
)
pluck_depth(x)
x |> map_int(pluck_depth)

[Package purrr version 1.0.1 Index]