rdunif {purrr}	R Documentation

Generate random sample from a discrete uniform distribution

Description

This function was deprecated in purrr 1.0.0 because it's not related to the core purpose of purrr.

Usage

rdunif(n, b, a = 1)

Arguments

n
Number of samples to draw.

a, b
Range of the distribution (inclusive).

Examples

These examples are designed to work in R >= 4.1 so that we can take advantage of modern syntax like the base pipe (|>) and the function shorthand (\(x) x + 1). They might not work on the version of R that you're using.

table(rdunif(1e3, 10))
table(rdunif(1e3, 10, -5))

[Package purrr version 1.0.1 Index]