Convert an object into a mapper function

as_mapper is the powerhouse behind the varied function specifications that most purrr functions allow. It is an S3 generic. The default method forwards its arguments to rlang::as_function().

as_mapper(.f, ...)

# S3 method for character
as_mapper(.f, ..., .null, .default = NULL)

# S3 method for numeric
as_mapper(.f, ..., .null, .default = NULL)

# S3 method for list
as_mapper(.f, ..., .null, .default = NULL)

Arguments

.f
A function, formula, or vector (not necessarily atomic).

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions. Note that formula functions conceptually take dots (that's why you can use ..1 etc). They silently ignore additional arguments that are not used in the formula expression.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to methods.

.default, .null
Optional additional argument for extractor functions (i.e. when .f is character, integer, or list). Returned when value is absent (does not exist) or empty (has length 0). .null is deprecated; please use .default instead.

Examples

as_mapper(~ . + 1)

#> <lambda>
#> function (..., .x = ..1, .y = ..2, . = ..1) 
#> . + 1
#> <environment: 0x7fc439edeab0>
#> attr(,"class")
#> [1] "rlang_lambda_function" "function"             

as_mapper(1)

#> function (x, ...) 
#> pluck(x, 1, .default = NULL)
#> <environment: 0x7fc439d9a460>

as_mapper(c("a", "b", "c"))

#> function (x, ...) 
#> pluck(x, "a", "b", "c", .default = NULL)
#> <environment: 0x7fc439d17bd8>

# Equivalent to function(x) x[["a"]][["b"]][["c"]]

as_mapper(list(1, "a", 2))

#> function (x, ...) 
#> pluck(x, 1, "a", 2, .default = NULL)
#> <environment: 0x7fc439ca1740>

# Equivalent to function(x) x[[1]][["a"]][[2]]

as_mapper(list(1, attr_getter("a")))

#> function (x, ...) 
#> pluck(x, 1, function (x) 
#> attr(x, attr, exact = TRUE), .default = NULL)
#> <environment: 0x7fc439c2ace8>

# Equivalent to function(x) attr(x[[1]], "a")

as_mapper(c("a", "b", "c"), .default = NA)

#> function (x, ...) 
#> pluck(x, "a", "b", "c", .default = NA)
#> <environment: 0x7fc439b86428>