Find the value or position of the first match

Find the value or position of the first match

Usage

detect(
  .x,
  .f,
  ...,
  .dir = c("forward", "backward"),
  .right = NULL,
  .default = NULL
)

detect_index(.x, .f, ..., .dir = c("forward", "backward"), .right = NULL)

Arguments

.x

A list or atomic vector.

.f

A function, formula, or vector (not necessarily atomic).

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions. Note that formula functions conceptually take dots (that's why you can use ..1 etc). They silently ignore additional arguments that are not used in the formula expression.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. If a component is not present, the value of .default will be returned.

...

Additional arguments passed on to the mapped function.

.dir

If "forward", the default, starts at the beginning of the vector and move towards the end; if "backward", starts at the end of the vector and moves towards the beginning.

.right

Please use .dir instead.

.default

The value returned when nothing is detected.

Value

detect the value of the first item that matches the predicate; detect_index the position of the matching item. If not found, detect returns NULL and detect_index

returns 0.

See also

Examples

is_even <- function(x) x %% 2 == 0

3:10 %>% detect(is_even)
#> [1] 4
3:10 %>% detect_index(is_even)
#> [1] 2

3:10 %>% detect(is_even, .dir = "backward")
#> [1] 10
3:10 %>% detect_index(is_even, .dir = "backward")
#> [1] 8

# Since `.f` is passed to as_mapper(), you can supply a
# lambda-formula or a pluck object:
x <- list(
  list(1, foo = FALSE),
  list(2, foo = TRUE),
  list(3, foo = TRUE)
)

detect(x, "foo")
#> [[1]]
#> [1] 2
#> 
#> $foo
#> [1] TRUE
#> 
detect_index(x, "foo")
#> [1] 2

# If you need to find all values, use keep():
keep(x, "foo")
#> [[1]]
#> [[1]][[1]]
#> [1] 2
#> 
#> [[1]]$foo
#> [1] TRUE
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 3
#> 
#> [[2]]$foo
#> [1] TRUE
#> 
#> 

# If you need to find all positions, use map_lgl():
which(map_lgl(x, "foo"))
#> [1] 2 3