every {purrr}	R Documentation

Do every or some elements of a list satisfy a predicate?

Description

Do every or some elements of a list satisfy a predicate?

Usage

every(.x, .p, ...)

some(.x, .p, ...)

Arguments

.x
A list or atomic vector.

.p
A predicate function to apply on each element of .x. some() returns TRUE when .p is TRUE for at least one element. every() returns TRUE when .p is TRUE for all elements.

...
Additional arguments passed on to .p.

Value

A logical vector of length 1.

Examples

y <- list(0:10, 5.5)
y %>% every(is.numeric)
y %>% every(is.integer)

[Package purrr version 0.3.2 Index]