Do every or some elements of a list satisfy a predicate?

Do every or some elements of a list satisfy a predicate?

every(.x, .p, ...)

some(.x, .p, ...)

Arguments

.x
A list or atomic vector.

.p
A single predicate function, a formula describing such a predicate function, or a logical vector of the same length as .x. Alternatively, if the elements of .x are themselves lists of objects, a string indicating the name of a logical element in the inner lists. Only those elements where .p evaluates to TRUE will be modified.

...
Additional arguments passed on to .f.

Value

A logical vector of length 1.

Examples

x <- list(0, 1, TRUE)
x %>% every(identity)

#> [1] TRUE

x %>% some(identity)

#> [1] TRUE

y <- list(0:10, 5.5)
y %>% every(is.numeric)

#> [1] TRUE

y %>% every(is.integer)

#> [1] FALSE