Apply a function to each element of a vector, and its index

imap_xxx(x, ...), an indexed map, is short hand for map2(x, names(x), ...) if x has names, or map2(x, seq_along(x), ...) if it does not. This is useful if you need to compute on both the value and the position of an element.

imap(.x, .f, ...)

imap_lgl(.x, .f, ...)

imap_chr(.x, .f, ...)

imap_int(.x, .f, ...)

imap_dbl(.x, .f, ...)

imap_dfr(.x, .f, ..., .id = NULL)

imap_dfc(.x, .f, ..., .id = NULL)

iwalk(.x, .f, ...)

Arguments

.x
A list or atomic vector.

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get_attr() to extract named attributes. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f.

.id
If not NULL a variable with this name will be created giving either the name or the index of the data frame.

Value

A vector the same length as .x.

See also

Other map variants: invoke, lmap, map2, map, modify

Examples

# Note that when using the formula shortcut, the first argument
# is the value, and the second is the position
imap_chr(sample(10), ~ paste0(.y, ": ", .x))

#>  [1] "1: 8"  "2: 7"  "3: 9"  "4: 2"  "5: 1"  "6: 10" "7: 5"  "8: 6"  "9: 3" 
#> [10] "10: 4"

iwalk(mtcars, ~ cat(.y, ": ", median(.x), "\n", sep = ""))

#> mpg: 19.2
#> cyl: 6
#> disp: 196.3
#> hp: 123
#> drat: 3.695
#> wt: 3.325
#> qsec: 17.71
#> vs: 0
#> am: 0
#> gear: 4
#> carb: 2