keep_at {purrr}	R Documentation

Keep/discard elements based on their name/position

Description

Keep/discard elements based on their name/position

Usage

keep_at(x, at)

discard_at(x, at)

Arguments

x
A list or atomic vector.

at
A logical, integer, or character vector giving the elements to select. Alternatively, a function that takes a vector of names, and returns a logical, integer, or character vector of elements to select.

: if the tidyselect package is installed, you can use vars() and tidyselect helpers to select elements.

Examples

These examples are designed to work in R >= 4.1 so that we can take advantage of modern syntax like the base pipe (|>) and the function shorthand (\(x) x + 1). They might not work on the version of R that you're using.

x <- c(a = 1, b = 2, cat = 10, dog = 15, elephant = 5, e = 10)
x 
x 

# Can also use a function
x 
x 

See Also

keep()/discard() to keep/discard elements by value.

[Package purrr version 1.0.1 Index]