Map over multiple inputs simultaneously.

These functions are variants of map() iterate over multiple arguments in parallel. map2() and walk2() are specialised for the two argument case; pmap() and pwalk() allow you to provide any number of arguments in a list.

map2(.x, .y, .f, ...)

map2_lgl(.x, .y, .f, ...)

map2_int(.x, .y, .f, ...)

map2_dbl(.x, .y, .f, ...)

map2_chr(.x, .y, .f, ...)

map2_dfr(.x, .y, .f, ..., .id = NULL)

map2_dfc(.x, .y, .f, ...)

walk2(.x, .y, .f, ...)

pmap(.l, .f, ...)

pmap_lgl(.l, .f, ...)

pmap_int(.l, .f, ...)

pmap_dbl(.l, .f, ...)

pmap_chr(.l, .f, ...)

pmap_dfr(.l, .f, ..., .id = NULL)

pmap_dfc(.l, .f, ...)

pwalk(.l, .f, ...)

Arguments

.x, .y
Vectors of the same length. A vector of length 1 will be recycled.

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get_attr() to extract named attributes. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f.

.id
If not NULL a variable with this name will be created giving either the name or the index of the data frame.

.l
A list of lists. The length of .l determines the number of arguments that .f will be called with. List names will be used if present.

Value

An atomic vector, list, or data frame, depending on the suffix. Atomic vectors and lists will be named if .x or the first element of .l is named.

If all input is length 0, the output will be length 0. If any input is length 1, it will be recycled to the length of the longest.

Details

Note that arguments to be vectorised over come before the .f, and arguments that are supplied to every call come after .f.

See also

Other map variants: imap, invoke, lmap, map, modify

Examples

x <- list(1, 10, 100)
y <- list(1, 2, 3)
z <- list(5, 50, 500)

map2(x, y, ~ .x + .y)

#> [[1]]
#> [1] 2
#> 
#> [[2]]
#> [1] 12
#> 
#> [[3]]
#> [1] 103
#> 

# Or just
map2(x, y, `+`)

#> [[1]]
#> [1] 2
#> 
#> [[2]]
#> [1] 12
#> 
#> [[3]]
#> [1] 103
#> 

# Split into pieces, fit model to each piece, then predict
by_cyl <- mtcars %>% split(.$cyl)
mods <- by_cyl %>% map(~ lm(mpg ~ wt, data = .))
map2(mods, by_cyl, predict)

#> $`4`
#>     Datsun 710      Merc 240D       Merc 230       Fiat 128    Honda Civic 
#>       26.47010       21.55719       21.78307       27.14774       30.45125 
#> Toyota Corolla  Toyota Corona      Fiat X1-9  Porsche 914-2   Lotus Europa 
#>       29.20890       25.65128       28.64420       27.48656       31.02725 
#>     Volvo 142E 
#>       23.87247 
#> 
#> $`6`
#>      Mazda RX4  Mazda RX4 Wag Hornet 4 Drive        Valiant       Merc 280 
#>       21.12497       20.41604       19.47080       18.78968       18.84528 
#>      Merc 280C   Ferrari Dino 
#>       18.84528       20.70795 
#> 
#> $`8`
#>   Hornet Sportabout          Duster 360          Merc 450SE          Merc 450SL 
#>            16.32604            16.04103            14.94481            15.69024 
#>         Merc 450SLC  Cadillac Fleetwood Lincoln Continental   Chrysler Imperial 
#>            15.58061            12.35773            11.97625            12.14945 
#>    Dodge Challenger         AMC Javelin          Camaro Z28    Pontiac Firebird 
#>            16.15065            16.33700            15.44907            15.43811 
#>      Ford Pantera L       Maserati Bora 
#>            16.91800            16.04103 
#> 

pmap(list(x, y, z), sum)

#> [[1]]
#> [1] 7
#> 
#> [[2]]
#> [1] 62
#> 
#> [[3]]
#> [1] 603
#> 

# Matching arguments by position
pmap(list(x, y, z), function(a, b ,c) a / (b + c))

#> [[1]]
#> [1] 0.1666667
#> 
#> [[2]]
#> [1] 0.1923077
#> 
#> [[3]]
#> [1] 0.1988072
#> 

# Matching arguments by name
l <- list(a = x, b = y, c = z)
pmap(l, function(c, b, a) a / (b + c))

#> [[1]]
#> [1] 0.1666667
#> 
#> [[2]]
#> [1] 0.1923077
#> 
#> [[3]]
#> [1] 0.1988072
#> 

# Vectorizing a function over multiple arguments
df <- data.frame(
  x = c("apple", "banana", "cherry"),
  pattern = c("p", "n", "h"),
  replacement = c("x", "f", "q"),
  stringsAsFactors = FALSE
  )
pmap(df, gsub)

#> [[1]]
#> [1] "axxle"
#> 
#> [[2]]
#> [1] "bafafa"
#> 
#> [[3]]
#> [1] "cqerry"
#> 

pmap_chr(df, gsub)

#> [1] "axxle"  "bafafa" "cqerry"

## Use `...` to absorb unused components of input list .l
df <- data.frame(
  x = 1:3 + 0.1,
  y = 3:1 - 0.1,
  z = letters[1:3]
)
plus <- function(x, y) x + y
## Not run: ------------------------------------
# ## this won't work
# pmap(df, plus)
## ---------------------------------------------
## but this will
plus2 <- function(x, y, ...) x + y
pmap_dbl(df, plus2)

#> [1] 4 4 4