prepend {purrr}	R Documentation

Prepend a vector

Description

This function was deprecated in purrr 1.0.0 because it's not related to the core purpose of purrr.

This is a companion to append() to help merging two lists or atomic vectors. prepend() is a clearer semantic signal than c() that a vector is to be merged at the beginning of another, especially in a pipe chain.

Usage

prepend(x, values, before = NULL)

Arguments

x
the vector to be modified.

values
to be included in the modified vector.

before
a subscript, before which the values are to be appended. If NULL, values will be appended at the beginning even for x of length 0.

Value

A merged vector.

Examples

x <- as.list(1:3)

x |> append("a")
x |> prepend("a")
x |> prepend(list("a", "b"), before = 3)
prepend(list(), x)

[Package purrr version 1.0.2 Index]