Set names in a vector

This is a snake case wrapper for stats::setNames(), with tweaked defaults, and stricter argument checking.

set_names(x, nm = x)

Arguments

x
Vector to name

nm
Vector of names, the same length as x.

If x already has names, you can provide a function or formula to transform the existing names.

Value

.x with the names attribute set.

Examples

set_names(1:4, c("a", "b", "c", "d"))

#> a b c d 
#> 1 2 3 4 

# If the second argument is ommitted a vector is named with itself
set_names(letters[1:5])

#>   a   b   c   d   e 
#> "a" "b" "c" "d" "e" 

# Alternatively you can supply a function
set_names(1:10, ~ letters[seq_along(.)])

#>  a  b  c  d  e  f  g  h  i  j 
#>  1  2  3  4  5  6  7  8  9 10 

set_names(head(mtcars), toupper)

#>                    MPG CYL DISP  HP DRAT    WT  QSEC VS AM GEAR CARB
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1