update_list {purrr}	R Documentation

Update a list with formulas

Description

update_list() was deprecated in purrr 1.0.0, because we no longer believe that functions that use NSE are a good fit for purrr.

update_list() handles formulas and quosures that can refer to values existing within the input list. This function is deprecated because we no longer believe that functions that use tidy evaluation are a good fit for purrr.

Usage

update_list(.x, ...)

Arguments

.x
List to modify.

...
New values of a list. Use zap() to remove values.

These values should be either all named or all unnamed. When inputs are all named, they are matched to .x by name. When they are all unnamed, they are matched by position.

Dynamic dots are supported. In particular, if your replacement values are stored in a list, you can splice that in with !!!.

[Package purrr version 1.0.4 Index]